Add files via upload
TASK 2 - Unemployment Analysis in India

Unemployment is measured by the unemployment rate which is the number of people who are unemployed as a percentage of the total labour force. We have seen a sharp increase in the unemployment rate during Covid 19, so analyzing the unemployment rate can be good data science project

The primary objective of this analysis is to conduct a comprehensive examination of regional unemployment rates in India without employing machine learning techniques. The dataset provides insights into unemployment dynamics, featuring key indicators such as the unemployment rate, estimated employment figures, and labor participation rates across different regions and time periods.